+++ sequence/components/__init__.py
@@ -1,5 +1,9 @@
+import importlib
 __all__ = ['beam_splitter', 'bsm', 'detector', 'interferometer', 'light_source', 'memory', 'optical_channel', 'photon',
            'spdc_lens', 'switch', 'circuit', 'transmon', 'transducer']
+
+for module in __all__:
+    importlib.import_module(f'.{module}', package='sequence.components')
 
 def __dir__():
     return sorted(__all__)
