+++ sequence/components/memory/__init__.py
@@ -5,7 +5,5 @@ Memories will attempt to send photons through the `send_qubit` interface of node
 Photons should be routed to a BSM device for entanglement generation, or through optical hardware for purification and swapping.
 """
 
-from .memory_base import Memory
-from .memory_array import MemoryArray
-from .absorptive_memory import AbsorptiveMemory
+from .base import Memory, MemoryArray, AbsorptiveMemory
 from .random_coherence_memory import MemoryWithRandomCoherenceTime
